backend/src/graphql/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
click==8.2.1
exceptiongroup==1.3.0
fastapi==0.104.1
graphql-core==3.2.6
h11==0.16.0
httptools==0.6.4
idna==3.10
libcst==1.7.0
markdown-it-py==3.0.0
mdurl==0.1.2
packaging==25.0
pydantic==2.11.5
pydantic_core==2.33.2
Pygments==2.19.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.6
PyYAML==6.0.2
rich==14.0.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
starlette==0.46.2
strawberry-graphql==0.213.0
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.14.0
uvicorn==0.24.0
uvloop==0.21.0
watchfiles==1.0.5
websockets==15.0.1
sqlalchemy==2.0.23
